Implemented Hydralisk
The Hydralisk should be done now. It registeres with the Arbiter and
recieves the overmind controlling it. It then passes all events on to
the overmind.

The Arbiter will create team names based on the names of the team
members and create an OverMind for the team. This allows multiple teams
of OverMinds to fight each other.

Overminds have stubs for the event methods it will need to handle.